Filter Estimated Values

ParameterEstimation.filter_solutions — Function

filter_solutions(results::Vector{EstimationResult},
				identifiability_result::IdentifiabilityData,
				model::ModelingToolkit.ODESystem,
				inputs::Vector{ModelingToolkit.Equation},
				data_sample::AbstractDict{Any, Vector{T}} = Dict{Any, Vector{T}}();
				solver = Tsit5(),
				topk = 1) where {T <: Float}

Filter estimation results stored in results vector based on ODE solving and checking against the sample. In addition, takes into account global and local identifiability of parameters when filtering.

Arguments

results::Vector{EstimationResult}: the vector of estimation results.

identifiability_result::IdentifiabilityData: the result of identifiability analysis.

model::ModelingToolkit.ODESystem: the ODE system.

inputs::Vector{ModelingToolkit.Equation}: the inputs of the ODE system.

data_sample::AbstractDict{Any, Vector{T}} = Dict{Any, Vector{T}}(): the data sample used for estimation (same functions as measured_quantities).

The keys of the dictionary are the measured quantities
															and the values are the corresponding data samples.

time_interval::Vector{T} = Vector{T}(): the time interval of the ODE system.

topk = 1: (optional) the number of best estimates to return.

Returns

EstimationResult: the best estimate (if topk = 1) or the vector of best estimates (if topk > 1).

ParameterEstimation.solve_ode — Function

solve_ode(model, estimate::EstimationResult, data_sample; solver = Tsit5(),
		  return_ode = false)

Solves the ODE system model with the parameters and initial conditions given by estimate. Compute the error between the solution and the data sample. The error is recorded in the EstimationResult.

Arguments

model: the ODE system to be solved.

estimate::EstimationResult: the parameters and initial conditions of the ODE system.

data_sample: the data sample used for estimation (same functions as measured_quantities).

 The keys of the dictionary are the measured quantities
			 and the values are the corresponding data samples.

solver = Tsit5(): (optional) the solver used to solve the ODE system, see DifferentialEquations for available solvers.

return_ode = false: (optional) whether to return the ODE solution.

Returns

odesolution: the solution of the ODE system (if `returnodeis set totrue`).

EstimationResult: the estimated parameters and initial conditions of the model.

ParameterEstimation.solve_ode! — Function

solve_ode!(model, estimates::Vector{EstimationResult},  inputs::Vector{Equation},data_sample; solver = Tsit5())

Run solve_ode for multiple estimates and store the results (error between solution and sample) in each estimate. This is done in-place.