Identifiability Information

ParameterEstimation.check_identifiability — Function

check_identifiability(ode::ModelingToolkit.ODESystem; measured_quantities = Array{ModelingToolkit.Equation}[],
                      infolevel = 0)

Check identifiability of parameters in the ODE system ode using the algorithm described in [1]. The function returns a ParameterEstimation.IdentifiabilityData object that contains the results of the identifiability analysis.

Arguments

- `ode::ModelingToolkit.ODESystem`: The ODE system to be analyzed
- `measured_quantities = Array{ModelingToolkit.Equation}[]`: A list of equations
    that define the measured quantities. If not provided, the outputs of the ODE
    system will be used.
- `infolevel::Int`: The level of information to be printed during the analysis.

References

[1] - Global Identifiability of Differential Models (Communications on Pure and Applied Mathematics, Volume 73, Issue 9, Pages 1831-1879, 2020.), https://onlinelibrary.wiley.com/doi/abs/10.1002/cpa.21921

[2] - SIAN: software for structural identifiability analysis of ODE models (Bioinformatics, Volume 35, Issue 16, Pages 2873–2874, 2019), https://academic.oup.com/bioinformatics/article/35/16/2873/5096881

[3] - https://github.com/alexeyovchinnikov/SIAN-Julia

ParameterEstimation.IdentifiabilityData — Type

IdentifiabilityData

A struct that contains the data from identifiability analysis. This is used for parameter estimation.

Fields

polynomial_system::Vector{SIAN.Nemo.fmpq_mpoly}: The polynomial system.

polynomial_system_to_solve::PolySystem: The polynomial system with derivatives substitutited and ready to be solved.

denominator::SIAN.Nemo.fmpq_mpoly: The denominator of the polynomial system.

variables::Vector{SIAN.Nemo.fmpq_mpoly}: The variables of the polynomial system.

substitutions::Vector{Vector}: The substitutions used to assess identifiability.

identifiability_nemo::Any: The identifiability data from SIAN in Nemo data type.

identifiability::Dict{Any, Any}: The identifiability data from SIAN in HomotopyContinuation compatible data type.

basis::Vector{SIAN.Nemo.fmpq_mpoly}: The transcendence basis of the polynomial system.

transcendence_basis_subs::Vector{SIAN.Nemo.RingElem}: The transcendence basis substitutions of the polynomial system.

weights::Dict{SIAN.Nemo.fmpq_mpoly, Int64}: The weights of the variables used by SIAN to assess GroebnerBasis.