Estimate

ParameterEstimation.estimate — Function

estimate(model::ModelingToolkit.ODESystem,
		measured_quantities::Vector{ModelingToolkit.Equation},
		data_sample::Dict{Any, Vector{T}} = Dict{Any, Vector{T}}();
		at_time::T = 0.0, method = :homotopy, solver = Tsit5(),
		degree_range = nothing, real_tol = 1e-12,
		threaded = Threads.nthreads() > 1) where {T <: Float}

Run estimation over a range of interpolation degrees. Return the best estimate according to a heuristic: - the best estimate is the one with the smallest error between sample data and ODE solution with current parameters (estimates);

Arguments

model::ModelingToolkit.ODESystem: the ODE model;

measured_quantities::Vector{ModelingToolkit.Equation}: the measured quantities (output functions that were sampled experimentally);

data_sample::Dict{Any, Vector{T}} = Dict{Any, Vector{T}}(): the data sample, a dictionary with keys being the measured quantities and

values being the corresponding data. Must include the time vector;

at_time::T = 0.0: the time used for derivative computation;

report_time = nothing: specify a time T, at which the initial conditions (state variables) will be estimated. If "nothing", use the leftmost time.

method = :homotopy: the method used for polynomial system solving. Can be one of :homotopy (recommended) or :msolve;

solver: the ODE solver used for ODE solution computation (default: Vern9());

interpolators = nothing: the set of interpolators to be used. See examples. If nothing, a default is used which includes AAA, FLoater-Hormann, and Fourier interpolations;

real_tol = 1e-14: the tolerance used for real root finding;

threaded = Threads.nthreads() > 1: whether to use multiple threads for computation (determined automatically).

Returns

result::Vector{EstimationResult}: the result of the estimation, a vector of EstimationResult objects.

Missing docstring.

Missing docstring for ParameterEstimation.estimate. Check Documenter's build log for details.

ParameterEstimation.EstimationResult — Type

EstimationResult

A container for the results of an estimation. Contains the estimated parameters and initial conditions (state values at a given time), the degree of the rational interpolation used, the error between the estimated ODE solution and the sample data, and the return code.

Fields

parameters::OrderedDict: The estimated parameters.

states::OrderedDict: The estimated initial conditions.

degree::Int64: The degree of the rational interpolation used.

at_time::Float64: The time at which the initial conditions are estimated.

err::Union{Nothing, Float64}: The error between the estimated ODE solution and the sample data.

interpolants::Union{Nothing, Dict{Any, Interpolant}}: The rational interpolants used to estimate the parameters and initial conditions.

return_code::Any: The return code of the estimation.

datasize::Int64: The number of data points used in the estimation.

report_time::Any: The time at which the initial conditions are reported (usually the leftmost point in the time span).