Rational Interpolation utils

ParameterEstimation.interpolate — Function

interpolate(identifiability_result, data_sample,
            measured_quantities; interpolation_degree::Int = 1,
            diff_order::Int = 1, at_t::Float = 0.0,
            method::Symbol = :homotopy)

This function performs the key step in parameter estimation.

It interpolates the data in `data_sample` and computes the `TaylorSeries` expansion.
These results are stored in the `Interpolant` object and are applied to the polynomial system in `identifiability_result`.

Arguments

identifiability_result: the result of the identifiability check.

data_sample: a dictionary of the data samples. The keys are the symbols of the measured quantities and the values are the data samples.

measured_quantities: the measured quantities (outputs as equations of the form y ~ x).

interpolation_degree::Int = 1: the degree of the numerator of the rational interpolation.

diff_order::Int = 1: the order of the derivative to be computed.

at_t::Float = 0.0: the time point where the Taylor series expansion is computed.

method::Symbol = :homotopy: the method used to solve the polynomial system. Can be one of :homotopy (recommended) and :msolve.

Returns

System: the polynomial system with the interpolated data applied. This system is compatible with HomotopyContinuation solving.

interpolate(time, sample, numer_degree::Int, diff_order::Int = 1, at_t::Float = 0.0)

This function performs a rational interpolation of the data sample at the points time with numerator degree numer_degree. It returns an Interpolant object that contains the interpolated function and its derivatives.

ParameterEstimation.rational_interpolation_coefficients — Function

rational_interpolation_coefficients(x, y, n)

Perform a rational interpolation of the data y at the points x with numerator degree n. This function only returns the coefficients of the numerator and denominator polynomials.

Arguments

x: the points where the data is sampled (e.g. time points).

y: the data sample.

n: the degree of the numerator.

Returns

c: the coefficients of the numerator polynomial.

d: the coefficients of the denominator polynomial.